vcmi: move bonuses to its own folder
