+++ lib/CTownHandler.h
@@ -19,7 +19,7 @@
 #include "IHandlerBase.h"
 #include "LogicalExpression.h"
 #include "battle/BattleHex.h"
-#include "HeroBonus.h"
+#include "bonuses/HeroBonus.h"
 #include "Point.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
