+++ lib/bonuses/HeroBonus.cpp
