+++ lib/bonuses/HeroBonus.h
